migrate all blocks except LargeShipYard to basic json model instead of TESR

+++ src/main/java/com/lulan/shincolle/client/render/block/RenderLargeShipyard.java
@@ -61,7 +61,7 @@ public class RenderLargeShipyard extends TileEntitySpecialRenderer<BasicTileEnti
 //        }
 //        else
 //        {
-        pitch += Math.PI * 0.5F;
+        pitch += (float) (Math.PI * 0.5F);
 //        }
 
         //render base
